+++ app/src/main/java/org/spsl/evtracker/domain/repository/ChargeEventQueries.kt
@@ -4,8 +4,8 @@ import kotlinx.coroutines.flow.Flow
 import org.spsl.evtracker.data.local.entity.ChargeEventEntity
 
 interface ChargeEventQueries {
-    fun observeForCar(carId: Int): Flow<List<ChargeEventEntity>>
-    suspend fun getInRange(carId: Int, from: Long, to: Long): List<ChargeEventEntity>
-    suspend fun getAllForCarSorted(carId: Int): List<ChargeEventEntity>
-    suspend fun getById(id: Int): ChargeEventEntity?
+    fun observeForCar(carId: Long): Flow<List<ChargeEventEntity>>
+    suspend fun getInRange(carId: Long, from: Long, to: Long): List<ChargeEventEntity>
+    suspend fun getAllForCarSorted(carId: Long): List<ChargeEventEntity>
+    suspend fun getById(id: Long): ChargeEventEntity?
 }
